Sieve Of Eratosthenes
This function produce prime numbers for a given range.

Initially we consider all numbers are prime..
Every complex number can be written as  multiple of some prime numbers..
Suppose we have found a number and it is prime
so it's multiples are not prime.
so we mark them as they are not prime..
again we took the next prime number and do the same..

Complexity: O(n(log n)(log log n))